requirements.txt
modal
selenium
numpy
scipy
scikit-image
torch
Pillow
transformers
